Primary shadow - cast shadow not working

+++ LubiEngine/DXBase.cpp
@@ -148,7 +148,7 @@ bool DXBase::Initialize(HWND hwnd, int width, int height, bool fullscreen, bool 
 	XMMATRIX world = XMMatrixIdentity();
 	XMStoreFloat4x4(&m_world, world);
 	
-	XMMATRIX projection = XMMatrixPerspectiveFovLH(XM_PIDIV4, (float)width / height, 0.1f, 1000.0f);
+	XMMATRIX projection = XMMatrixPerspectiveFovLH(XM_PIDIV4, (float)width / (float)height, 0.1f, 1000.0f);
 	XMStoreFloat4x4(&m_projection, projection);
 
 	XMMATRIX ortho = XMMatrixOrthographicLH((float)width, (float)height, 0.1f, 1000.0f);
